Add useCategoryLeaf, useCategoryBranch
