+++ packages/peregrine/lib/talons/CategoryTree/index.js
@@ -1,1 +1,3 @@
+export { useCategoryBranch } from './useCategoryBranch';
+export { useCategoryLeaf } from './useCategoryLeaf';
 export { useCategoryTree } from './useCategoryTree';
