+++ packages/venia-ui/lib/components/CategoryTree/categoryLeaf.js
@@ -1,5 +1,6 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { func, shape, string } from 'prop-types';
+import { useCategoryLeaf } from '@magento/peregrine/lib/talons/CategoryTree';
 
 import { mergeClasses } from '../../classify';
 import { Link, resourceUrl } from '../../drivers';
@@ -11,16 +12,14 @@ const Leaf = props => {
     const { category, onNavigate } = props;
     const { name, url_path } = category;
     const classes = mergeClasses(defaultClasses, props.classes);
-
-    const handleClick = useCallback(() => {
-        onNavigate();
-    }, [onNavigate]);
+    const { handleClick } = useCategoryLeaf({ onNavigate });
+    const destination = resourceUrl(`/${url_path}${suffix}`);
 
     return (
         <li className={classes.root}>
             <Link
                 className={classes.target}
-                to={resourceUrl(`/${url_path}${suffix}`)}
+                to={destination}
                 onClick={handleClick}
             >
                 <span className={classes.text}>{name}</span>
