+++ packages/venia-ui/lib/components/CategoryTree/categoryTree.js
@@ -9,6 +9,7 @@ import Leaf from './categoryLeaf';
 import defaultClasses from './categoryTree.css';
 
 const Tree = props => {
+    console.log(MENU_QUERY);
     const {
         categories,
         categoryId,
